requirements.txt
psycopg[binary]>=3.2,<3.4
python-dotenv==1.0.1
streamlit>=1.44,<2.0
pandas>=2.2,<3.0
matplotlib>=3.9,<4.0